+++ models/unit_tests.go
@@ -44,10 +44,6 @@ func MainTest(m *testing.M, pathToGiteaRoot string) {
 		fatalTestError("Error creating test engine: %v\n", err)
 	}
 
-	if err = InitIssueIndexer(); err != nil {
-		fatalTestError("Error InitIssueIndexer: %v\n", err)
-	}
-
 	setting.AppURL = "https://try.gitea.io/"
 	setting.RunUser = "runuser"
 	setting.SSH.Port = 3000
